feat: overlay 외부 클릭시 닫히게 구현
